rename new shape submodule to labelled
