+++ menpo/shape/labelled.py
